# specs/001-guest-lock-pins/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Guest Lock PIN Manager API
  description: |
    REST API for the Guest Lock PIN Manager Home Assistant addon.
    Manages calendar subscriptions, lock configurations, and PIN lifecycle.
  version: 1.0.0
  contact:
    name: Guest Lock Manager
servers:
  - url: /api
    description: Home Assistant ingress-proxied API
...
paths:
...
  /calendars:
    get:
      tags: [calendars]
      summary: List all calendar subscriptions
      operationId: listCalendars
      responses:
        '200':
          description: List of calendars
          content:
            application/json:
              schema:
                type: object
                properties:
                  calendars:
                    type: array
                    items:
                      $ref: '#/components/schemas/Calendar'
    post:
      tags: [calendars]
      summary: Subscribe to a new calendar
      operationId: createCalendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarCreate'
      responses:
        '201':
          description: Calendar created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Calendar URL already exists
  /calendars/{id}:
    parameters:
      - $ref: '#/components/parameters/CalendarId'
    get:
      tags: [calendars]
      summary: Get calendar details
      operationId: getCalendar
      responses:
        '200':
          description: Calendar details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [calendars]
      summary: Update calendar settings
      operationId: updateCalendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarUpdate'
      responses:
        '200':
          description: Calendar updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calendar'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [calendars]
      summary: Unsubscribe from calendar
      operationId: deleteCalendar
      responses:
        '204':
          description: Calendar deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /calendars/{id}/sync:
    parameters:
      - $ref: '#/components/parameters/CalendarId'
    post:
      tags: [calendars]
      summary: Trigger immediate calendar sync
      operationId: syncCalendar
      responses:
        '202':
          description: Sync started
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Sync initiated"
        '404':
          $ref: '#/components/responses/NotFound'
  /calendars/{id}/locks:
    parameters:
      - $ref: '#/components/parameters/CalendarId'
    get:
      tags: [calendars]
      summary: Get locks assigned to calendar
      operationId: getCalendarLocks
      responses:
        '200':
          description: Assigned locks
          content:
            application/json:
              schema:
                type: object
                properties:
                  locks:
                    type: array
                    items:
                      $ref: '#/components/schemas/LockSummary'
    put:
      tags: [calendars]
      summary: Update lock assignments for calendar
      operationId: updateCalendarLocks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lock_ids]
              properties:
                lock_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Locks assigned
        '404':
          $ref: '#/components/responses/NotFound'

  # ============== LOCKS ==============
...
components:
  parameters:
    CalendarId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: "not_found"
        message:
          type: string
          example: "Calendar not found"
        details:
          type: object
          additionalProperties: true
    Calendar:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
          format: uri
        sync_interval_min:
          type: integer
        last_sync_at:
          type: string
          format: date-time
          nullable: true
        sync_status:
          type: string
          enum: [pending, syncing, success, error]
        sync_error:
          type: string
          nullable: true
        enabled:
          type: boolean
        lock_count:
          type: integer
        pin_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CalendarCreate:
      type: object
      required: [name, url]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        url:
          type: string
          format: uri
        sync_interval_min:
          type: integer
          minimum: 5
          default: 15
        lock_ids:
          type: array
          items:
            type: string
            format: uuid
    CalendarUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        sync_interval_min:
          type: integer
          minimum: 5
        enabled:
          type: boolean
...
    LockSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        online:
          type: boolean
...